Fix: Recorder: Populate input device list when no devices are specified
ydotool has automatically provided a default device list when the user
did not specify one for a while. However, when the cxxopts library was
updated in
https://github.com/ReimuNotMoe/ydotool/commit/f4e91671187feb32c0564010f75a79e52539b0bd
the behavior of the default flag value function changed; now because the
lambda for the devices flag emits an empty string (""), the device list
is no longer empty. So this change adds an additional check to check if
the only value in the device list returned by cxxopts is just a single
empty string; in this case go ahead and populate a default device list
using the old method.

NOTE: It is possible that the user may now specify --devices "", and
this change will cause the defaults to be used anyway.

Signed-off-by: Brendan Higgins <bhig93@gmail.com>